standardize services link pattern
