+++ BFieldTest/src/SConscript
@@ -12,7 +12,7 @@ helper=mu2e_helper(env);
 rootlibs = env['ROOTLIBS']
 
 helper.make_plugins( [
-        'mu2e_SeedService_SeedService_service',
+        'mu2e_SeedService',
         'mu2e_GeometryService',
         'mu2e_BFieldGeom',
         'mu2e_Mu2eInterfaces',
