+++ CRVReco/src/SConscript
@@ -29,12 +29,11 @@ mainlib = helper.make_mainlib ( [ 'CLHEP',
                               )
 
 helper.make_plugins( [ mainlib,
-                       'mu2e_SeedService_SeedService_service',
+                       'mu2e_SeedService',
                        'mu2e_Mu2eUtilities',
                        'mu2e_GeometryService',
                        'mu2e_ConditionsService',
                        'mu2e_DAQConditions',
-                       'mu2e_GlobalConstantsService_GlobalConstantsService_service',
                        'mu2e_GlobalConstantsService',
                        'mu2e_RecoDataProducts',
                        'mu2e_CosmicRayShieldGeom',
@@ -47,7 +46,7 @@ helper.make_plugins( [ mainlib,
                        'art_Framework_Services_Registry',
                        'art_root_io_tfile_support',
                        'art_root_io_TFileService',
-                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Framework_Services_Optional_RandomNumberGenerator',
                        'art_Persistency_Common',
                        'art_Persistency_Provenance',
                        'art_Utilities',
