+++ CaloDiag/src/SConscript
@@ -38,7 +38,6 @@ helper.make_plugins( [ mainlib,
                        'mu2e_StoppingTargetGeom',
                        'mu2e_TrackerGeom',
                        'mu2e_TrackerConditions',
-                       'mu2e_GlobalConstantsService_GlobalConstantsService_service',
                        'mu2e_GlobalConstantsService',
                        'mu2e_ConfigTools',
                        'mu2e_Mu2eInterfaces',
@@ -51,7 +50,7 @@ helper.make_plugins( [ mainlib,
                        'art_Framework_Services_Registry',
                        'art_root_io_tfile_support',
                        'art_root_io_TFileService',
-                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Framework_Services_Optional_RandomNumberGenerator',
                        'art_Persistency_Common',
                        'art_Persistency_Provenance',
                        'art_Utilities',
