+++ CaloReco/src/SConscript
@@ -26,7 +26,7 @@ mainlib = helper.make_mainlib ( [ 'mu2e_ConditionsService',
                                   'art_Persistency_Common',
                                   'art_Persistency_Provenance',
                                   'canvas',
-                                  'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                                  'art_Framework_Services_Optional_RandomNumberGenerator',
                                   'art_root_io_TFileService',
                                   'art_Framework_Services_Registry',
                                   'art_root_io_tfile_support',
@@ -47,7 +47,7 @@ helper.make_plugins( [ mainlib,
                        'mu2e_Mu2eUtilities',
                        'mu2e_ConditionsService',
                        'mu2e_GeometryService',
-                       'mu2e_SeedService_SeedService_service',
+                       'mu2e_SeedService',
                        'mu2e_DAQConditions',
                        'mu2e_CalorimeterGeom',
                        'mu2e_RecoDataProducts',
@@ -59,7 +59,7 @@ helper.make_plugins( [ mainlib,
                        'art_Framework_Services_Registry',
                        'art_root_io_tfile_support',
                        'art_root_io_TFileService',
-                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Framework_Services_Optional_RandomNumberGenerator',
                        'art_Persistency_Common',
                        'art_Persistency_Provenance',
                        'art_Utilities',
