+++ CommonMC/src/SConscript
@@ -25,7 +25,7 @@ helper.make_plugins( [ mainlib,
   'mu2e_TrackerConditions',
   'mu2e_DAQConditions',
   'mu2e_GeometryService',
-  'mu2e_SeedService_SeedService_service',
+  'mu2e_SeedService',
   'mu2e_TrackerGeom',
   'mu2e_MCDataProducts',
   'mu2e_RecoDataProducts',
@@ -39,7 +39,7 @@ helper.make_plugins( [ mainlib,
   'art_Framework_Services_Registry',
   'art_root_io_tfile_support',
   'art_root_io_TFileService',
-  'art_Framework_Services_Optional_RandomNumberGenerator_service',
+  'art_Framework_Services_Optional_RandomNumberGenerator',
   'art_Persistency_Common',
   'art_Persistency_Provenance',
   'art_Utilities',
