+++ DAQ/src/SConscript
@@ -29,7 +29,7 @@ mainlib = helper.make_mainlib ( [ 'mu2e_ConditionsService',
                                   'DTCInterface',
                                   'art_Persistency_Provenance',
                                   'art_Persistency_Common',
-                                  'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                                  'art_Framework_Services_Optional_RandomNumberGenerator',
                                   'art_root_io_TFileService',
                                   'art_Framework_Services_Registry',
                                   'art_root_io_tfile_support',
@@ -54,7 +54,7 @@ helper.make_plugins( [ mainlib,
                        'mu2e_TrackerConditions',
                        'mu2e_CaloConditions',
                        'mu2e_GeometryService',
-                       'mu2e_SeedService_SeedService_service',
+                       'mu2e_SeedService',
                        'mu2e_BFieldGeom',
                        'mu2e_CalorimeterGeom',
                        'mu2e_TrackerGeom',
@@ -70,7 +70,7 @@ helper.make_plugins( [ mainlib,
                        'art_Framework_Services_Registry',
                        'art_root_io_tfile_support',
                        'art_root_io_TFileService',
-                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Framework_Services_Optional_RandomNumberGenerator',
                        'art_Persistency_Common',
                        'art_Persistency_Provenance',
                        'canvas',
