+++ DbService/src/SConscript
@@ -14,12 +14,18 @@ helper=mu2e_helper(env)
 mainlib = helper.make_mainlib ( [ 'mu2e_DbTables','mu2e_GeneralUtilities',
                                   'art_Framework_Core',
                                   'art_Framework_Services_Registry',
+                                  'art_Utilities',
+                                  'MF_MessageLogger',
+                                  'fhiclcpp',
+                                  'fhiclcpp_types',
+                                  'tbb',
                                   'CLHEP',
                                   'cetlib',
                                   'cetlib_except',
                                   'curl',
                                   'pq',
-                                  'pthread'
+                                  'pthread',
+                                  'boost_filesystem',
                                   ] )
 
 helper.make_plugins( [ mainlib,
@@ -29,7 +35,7 @@ helper.make_plugins( [ mainlib,
                        'art_Framework_Services_Registry',
                        'art_root_io_tfile_support',
                        'art_root_io_TFileService',
-                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Framework_Services_Optional_RandomNumberGenerator',
                        'art_Persistency_Common',
                        'art_Persistency_Provenance',
                        'art_Utilities',
