+++ EventGenerator/src/SConscript
@@ -17,7 +17,6 @@ crylib = os.environ['CRY_LIB']
 rootlibs  = env['ROOTLIBS']
 
 mainlib = helper.make_mainlib ( [
-                          'mu2e_GlobalConstantsService_GlobalConstantsService_service',
                           'mu2e_GlobalConstantsService',
                           'mu2e_GeometryService',
                           'mu2e_ConditionsService',
@@ -42,7 +41,7 @@ mainlib = helper.make_mainlib ( [
         'art_Persistency_Provenance',
         'art_Persistency_Common',
         'art_Framework_Services_Registry',
-        'art_Framework_Services_Optional_RandomNumberGenerator_service',
+        'art_Framework_Services_Optional_RandomNumberGenerator',
         'art_root_io_TFileService',
         'art_root_io_tfile_support',
         'art_Framework_Principal',
@@ -63,7 +62,7 @@ mainlib = helper.make_mainlib ( [
         )
 
 helper.make_plugins( [ mainlib,
-                       'mu2e_SeedService_SeedService_service',
+                       'mu2e_SeedService',
                        'mu2e_Mu2eUtilities',
                        'mu2e_ConditionsService',
                        'mu2e_GeometryService',
@@ -83,7 +82,7 @@ helper.make_plugins( [ mainlib,
                        'art_Framework_Services_Registry',
                        'art_root_io_tfile_support',
                        'art_root_io_TFileService',
-                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Framework_Services_Optional_RandomNumberGenerator',
                        'art_Persistency_Common',
                        'art_Persistency_Provenance',
                        'art_Utilities',
