+++ EventMixing/src/SConscript
@@ -20,7 +20,7 @@ mainlib = helper.make_mainlib ( [
     'art_Framework_Services_Registry',
     'art_root_io_tfile_support',
     'art_root_io_TFileService',
-    'art_Framework_Services_Optional_RandomNumberGenerator_service',
+    'art_Framework_Services_Optional_RandomNumberGenerator',
     'art_Framework_IO_ProductMix',
     'art_root_io',
     'art_root_io_detail',
@@ -41,7 +41,7 @@ mainlib = helper.make_mainlib ( [
 
 helper.make_plugins ( [ mainlib,
         'mu2e_Mu2eUtilities',
-        'mu2e_SeedService_SeedService_service',
+        'mu2e_SeedService',
         'mu2e_MCDataProducts',
         'mu2e_DataProducts',
         'mu2e_SimulationConditions',
@@ -51,7 +51,7 @@ helper.make_plugins ( [ mainlib,
         'art_Framework_Services_Registry',
         'art_root_io_tfile_support',
         'art_root_io_TFileService',
-        'art_Framework_Services_Optional_RandomNumberGenerator_service',
+        'art_Framework_Services_Optional_RandomNumberGenerator',
         'art_Framework_IO_ProductMix',
         'art_root_io',
         'art_root_io_detail',
