+++ ExtinctionMonitorFNAL/Digitization/src/SConscript
@@ -15,6 +15,7 @@ mainlib = helper.make_mainlib ( [
         'mu2e_GeometryService',
         'mu2e_ExtinctionMonitorFNAL_Geometry',
         'mu2e_ExtinctionMonitorFNAL_Utilities',
+        'mu2e_SeedService',
         'mu2e_MCDataProducts',
         'mu2e_RecoDataProducts',
         'mu2e_Mu2eInterfaces',
@@ -23,7 +24,7 @@ mainlib = helper.make_mainlib ( [
         'art_Persistency_Common',
         'art_Framework_Services_Registry',
         'art_root_io_tfile_support',
-        'art_Framework_Services_Optional_RandomNumberGenerator_service',
+        'art_Framework_Services_Optional_RandomNumberGenerator',
         'art_Framework_Principal',
         'art_Framework_Core',
         'art_Utilities',
@@ -43,6 +44,7 @@ helper.make_plugins( [ mainlib,
                        'mu2e_GlobalConstantsService',
                        'mu2e_SeedService_SeedService_service',
                        'mu2e_ExtinctionMonitorFNAL_Geometry',
+                       'mu2e_SeedService',
                        'mu2e_MCDataProducts',
                        'mu2e_RecoDataProducts',
                        'mu2e_Mu2eInterfaces',
@@ -51,7 +53,7 @@ helper.make_plugins( [ mainlib,
                        'art_Framework_Principal',
                        'art_Framework_Services_Registry',
                        'art_root_io_tfile_support',
-                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Framework_Services_Optional_RandomNumberGenerator',
                        'art_Persistency_Common',
                        'art_Persistency_Provenance',
                        'art_Utilities',
