+++ Mu2e/src/SConscript
@@ -16,7 +16,7 @@ helper.make_bin("mu2e", [
     'art_Persistency_Provenance',
     'art_Framework_Services_Registry',
     'art_root_io_tfile_support',
-    'art_Framework_Services_Optional_RandomNumberGenerator_service',
+    'art_Framework_Services_Optional_RandomNumberGenerator',
     'art_Utilities',
     'canvas',
     'MF_MessageLogger',
