+++ Mu2eKinKal/src/SConscript
@@ -23,7 +23,6 @@ mainlib = helper.make_mainlib([
   'mu2e_CalorimeterGeom',
   'mu2e_TrackerGeom',
   'mu2e_RecoDataProducts',
-  'mu2e_GlobalConstantsService_GlobalConstantsService_service',
   'mu2e_GlobalConstantsService',
   'mu2e_Mu2eInterfaces',
   'mu2e_DataProducts',
@@ -68,7 +67,6 @@ helper.make_plugins([mainlib,
   'mu2e_CalorimeterGeom',
   'mu2e_TrackerGeom',
   'mu2e_RecoDataProducts',
-  'mu2e_GlobalConstantsService_GlobalConstantsService_service',
   'mu2e_GlobalConstantsService',
   'mu2e_DbTables',
   'mu2e_Mu2eInterfaces',
@@ -81,7 +79,7 @@ helper.make_plugins([mainlib,
   'art_Framework_Services_Registry',
   'art_root_io_tfile_support',
   'art_root_io_TFileService',
-  'art_Framework_Services_Optional_RandomNumberGenerator_service',
+  'art_Framework_Services_Optional_RandomNumberGenerator',
   'art_Persistency_Common',
   'art_Persistency_Provenance',
   'art_Utilities',
