+++ Mu2eUtilities/src/SConscript
@@ -14,8 +14,8 @@ XERCESC_LIBS = [ 'xerces-c' ]
 
 rootlibs = env['ROOTLIBS']
 babarlibs = env['BABARLIBS']
-mainlib = helper.make_mainlib ( [ 'mu2e_ConditionsService_ConditionsService_service',
-                                  'mu2e_GeometryService_GeometryService_service',
+mainlib = helper.make_mainlib ( [ 'mu2e_ConditionsService',
+                                  'mu2e_GeometryService',
                                   'mu2e_GlobalConstantsService',
                                   'mu2e_StoppingTargetGeom',
                                   'mu2e_GeneralUtilities',
@@ -30,7 +30,7 @@ mainlib = helper.make_mainlib ( [ 'mu2e_ConditionsService_ConditionsService_serv
                                   'gsl',
                                   'art_Persistency_Common',
                                   'art_Persistency_Provenance',
-                                  'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                                  'art_Framework_Services_Optional_RandomNumberGenerator',
                                   'art_root_io_TFileService',
                                   'art_Framework_Services_Registry',
                                   'art_root_io_tfile_support',
