+++ ProditionsService/src/SConscript
@@ -12,13 +12,32 @@ Import('mu2e_helper')
 helper=mu2e_helper(env)
 
 mainlib = helper.make_mainlib ( [ 'mu2e_ConfigTools',
+                                  'mu2e_DAQConditions',
+                                  'mu2e_TrackerConditions',
+                                  'mu2e_CaloConditions',
+                                  'mu2e_SimulationConditions',
+                                  'mu2e_STMConditions',
+                                  'mu2e_CRVConditions',
+                                  'mu2e_Mu2eBTrk',
+                                  'mu2e_GeometryService',
+                                  'mu2e_GeomPrimitives',
                                   'mu2e_DbService',
                                   'mu2e_DbTables',
+                                  'mu2e_Mu2eInterfaces',
+                                  'mu2e_RecoDataProducts',
+                                  'mu2e_GeneralUtilities',
+                                  'mu2e_Mu2eUtilities',
                                   'art_Framework_Core',
                                   'art_Framework_Services_Registry',
+                                  'art_Utilities',
+                                  'MF_MessageLogger',
+                                  'canvas',
+                                  'fhiclcpp',
+                                  'fhiclcpp_types',
                                   'CLHEP',
                                   'cetlib',
                                   'cetlib_except',
+                                  'xerces-c',
                                   'curl',
                                   'pq'
                                   ] )
@@ -50,7 +69,7 @@ helper.make_plugins( [ mainlib,
                        'art_Framework_Services_Registry',
                        'art_root_io_tfile_support',
                        'art_root_io_TFileService',
-                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Framework_Services_Optional_RandomNumberGenerator',
                        'art_Persistency_Common',
                        'art_Persistency_Provenance',
                        'art_Utilities',
