+++ SeedService/src/SConscript
@@ -9,59 +9,48 @@ Import('mu2e_helper')
 
 helper=mu2e_helper(env)
 
-#
-# Fixme.
-# The link list below correctly builds the _service.so
-# To build the _module.so however, I need to add the _service.so
-# to the link list.  This triggers a rebuild of _service.so, which
-# then fails ...
-# 'mu2e_SeedService_SeedService_service',
-# Need to revise the build system to separately build modules and services.
-#
+mainlib = helper.make_mainlib ( [
+    'art_Framework_Core',
+    'art_Framework_IO_Sources',
+    'art_Framework_Principal',
+    'art_Persistency_Common',
+    'art_Persistency_Provenance',
+    'art_Framework_Services_Registry',
+    'art_root_io_tfile_support',
+    'art_Utilities',
+    'canvas',
+    'MF_MessageLogger',
+    'fhiclcpp',
+    'fhiclcpp_types',
+    'tbb',
+    'cetlib',
+    'cetlib_except',
+    'CLHEP',
+    'Core',
+    'boost_filesystem',
+    ] )
 
-helper.make_plugin( 'SeedService_service.cc',
-                    [
-                      'art_Framework_Core',
-                      'art_Framework_IO_Sources',
-                      'art_Framework_Principal',
-                      'art_Persistency_Common',
-                      'art_Persistency_Provenance',
-                      'art_Framework_Services_Registry',
-                      'art_root_io_tfile_support',
-                      'art_Utilities',
-                      'canvas',
-                      'MF_MessageLogger',
-                      'fhiclcpp',
-                      'fhiclcpp_types',
-                      'tbb',
-                      'cetlib',
-                      'cetlib_except',
-                      'CLHEP',
-                      'Core',
-                      'boost_filesystem',
-                      ] )
 
 helper.make_plugins( [
-                       'mu2e_SeedService_SeedService_service',
-                       'art_Framework_Core',
-                       'art_Framework_IO_Sources',
-                       'art_Framework_Principal',
-                       'art_Persistency_Common',
-                       'art_Persistency_Provenance',
-                       'art_Framework_Services_Registry',
-                       'art_root_io_tfile_support',
-                       'art_Utilities',
-                       'canvas',
-                       'MF_MessageLogger',
-                       'fhiclcpp',
-                       'cetlib',
-                       'cetlib_except',
-                       'CLHEP',
-                       'Core',
-                       'boost_filesystem',
-                       ],
-                     [ 'SeedService_service.cc', ],
-                     )
+    mainlib,
+    'art_Framework_Core',
+    'art_Framework_IO_Sources',
+    'art_Framework_Principal',
+    'art_Persistency_Common',
+    'art_Persistency_Provenance',
+    'art_Framework_Services_Registry',
+    'art_root_io_tfile_support',
+    'art_Utilities',
+    'canvas',
+    'MF_MessageLogger',
+    'fhiclcpp',
+    'cetlib',
+    'cetlib_except',
+    'CLHEP',
+    'Core',
+    'boost_filesystem',
+] )
+
 
 # This tells emacs to view this file in python mode.
 # Local Variables:
