+++ Sources/src/SConscript
@@ -14,7 +14,7 @@ mainlib = helper.make_mainlib(['mu2e_GlobalConstantsService',
                                'mu2e_Mu2eUtilities',
                                'mu2e_ConditionsService',
                                'mu2e_GeometryService',
-                               'mu2e_SeedService_SeedService_service',
+                               'mu2e_SeedService',
                                'mu2e_CalorimeterGeom',
                                'mu2e_ExtinctionMonitorFNAL_Geometry',
                                'mu2e_ProductionTargetGeom',
@@ -29,7 +29,7 @@ mainlib = helper.make_mainlib(['mu2e_GlobalConstantsService',
                                'art_Persistency_Provenance',
                                'art_Persistency_Common',
                                'art_Framework_Services_Registry',
-                               'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                               'art_Framework_Services_Optional_RandomNumberGenerator',
                                'art_root_io_TFileService',
                                'art_root_io_tfile_support',
                                'art_Framework_Principal',
@@ -70,7 +70,7 @@ helper.make_plugins( [ mainlib,
                        'Core',
                        'boost_filesystem',
                        'mu2e_GeometryService',
-                       'mu2e_SeedService_SeedService_service',
+                       'mu2e_SeedService',
                        'mu2e_CalorimeterGeom',
                        'mu2e_ExtinctionMonitorFNAL_Geometry',
                        'mu2e_ProductionTargetGeom',
