+++ TEveEventDisplay/src/SConscript
@@ -27,7 +27,6 @@ userlibs = [ rootlibs,
     'mu2e_Mu2eUtilities',
     'mu2e_MCDataProducts',
     'mu2e_RecoDataProducts',
-    'mu2e_GlobalConstantsService_GlobalConstantsService_service',
     'mu2e_GlobalConstantsService',
     'mu2e_GeometryService',
     'mu2e_ConditionsService',
@@ -43,7 +42,7 @@ userlibs = [ rootlibs,
     'art_Framework_Services_Registry',
     'art_root_io_tfile_support',
     'art_root_io_TFileService',
-    'art_Framework_Services_Optional_RandomNumberGenerator_service',
+    'art_Framework_Services_Optional_RandomNumberGenerator',
     'art_Persistency_Common',
     'art_Persistency_Provenance',
     'art_Utilities',
