+++ Trigger/src/SConscript
@@ -37,7 +37,6 @@ helper.make_plugins( [ mainlib,
                        'mu2e_ProtonBeamDumpGeom',
                        'mu2e_StoppingTargetGeom',
                        'mu2e_TrackerGeom',
-                       'mu2e_GlobalConstantsService_GlobalConstantsService_service',
                        'mu2e_GlobalConstantsService',
                        'mu2e_ConfigTools',
                        'mu2e_Mu2eInterfaces',
@@ -50,7 +49,7 @@ helper.make_plugins( [ mainlib,
                        'art_Framework_Services_Registry',
                        'art_root_io_tfile_support',
                        'art_root_io_TFileService',
-                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Framework_Services_Optional_RandomNumberGenerator',
 #                       'art_Framework_Services_System',
                        'art_Framework_Services_System_TriggerNamesService_service',
                        'art_Persistency_Common',
