+++ TrkDiag/src/SConscript
@@ -64,7 +64,7 @@ helper.make_plugins( [
   'mu2e_Mu2eBTrk',
   'mu2e_Mu2eUtilities',
   'mu2e_CaloCluster',
-  'mu2e_SeedService_SeedService_service',
+  'mu2e_SeedService',
   'mu2e_ConditionsService',
   'mu2e_GeometryService',
   'mu2e_BFieldGeom',
@@ -87,7 +87,7 @@ helper.make_plugins( [
   'art_Framework_Services_Registry',
   'art_root_io_tfile_support',
   'art_root_io_TFileService',
-  'art_Framework_Services_Optional_RandomNumberGenerator_service',
+  'art_Framework_Services_Optional_RandomNumberGenerator',
   'art_Framework_Services_System_TriggerNamesService_service',
   'art_Persistency_Common',
   'art_Persistency_Provenance',
