+++ UtilityModules/src/SConscript
@@ -16,7 +16,7 @@ common = [ 'mu2e_MCDataProducts',
            'mu2e_RecoDataProducts',
            'mu2e_ConfigTools',
            'mu2e_GeneralUtilities',
-           'mu2e_SeedService_SeedService_service',
+           'mu2e_SeedService',
            'art_Framework_Core',
            'art_Framework_IO_Sources',
            'art_Framework_Principal',
@@ -54,7 +54,7 @@ helper.make_plugins( [ mainlib,
                        'mu2e_RecoDataProducts',
                        'mu2e_ConfigTools',
                        'mu2e_GeneralUtilities',
-                       'mu2e_SeedService_SeedService_service',
+                       'mu2e_SeedService',
                        'art_Framework_Core',
                        'art_Framework_IO_Sources',
                        'art_Framework_Principal',
